Code enhancement: Fix GCC warnings
- fix conversion warnings
- fix missing-field-initializers warnings
- fix unused-parameter warnings
- fix deprecated-copy warnings
- add initializers

Fix #13122, close #13123

+++ PowerEditor/src/uchardet/CharDistribution.cpp
@@ -57,7 +57,7 @@ float CharDistributionAnalysis::GetConfidence(void)
 
   if (mTotalChars != mFreqChars)
   {
-    float r = mFreqChars / ((mTotalChars - mFreqChars) * mTypicalDistributionRatio);
+    float r = static_cast<float>(mFreqChars) / (static_cast<float>(mTotalChars - mFreqChars) * mTypicalDistributionRatio);
 
     if (r < SURE_YES)
       return r;
